# .scratch/mozhou-mvp/contracts/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "墨舟 MoZhou API",
    "version": "1.0.0",
    "description": "AI 小说写作平台 — 全站 19 组端点契约（UVSD Contract-Driven）"
  },
  "servers": [
    {
      "url": "/",
      "description": "同源部署"
    }
  ],
  "paths": {
...
    "/api/v1/chat": {
      "post": {
        "summary": "流式写作对话（SSE）",
        "operationId": "chatSend",
        "responses": {
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "200": {
            "description": "SSE stream (text/event-stream)"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sessionId": {
                    "type": "integer",
                    "nullable": true
                  },
                  "model": {
                    "type": "string",
                    "enum": [
                      "deepseek-v4-flash",
                      "glm-4.5-flash"
                    ]
                  },
                  "content": {
                    "type": "string"
                  },
                  "novelId": {
                    "type": "integer",
                    "nullable": true
                  },
                  "style": {
                    "type": "string",
                    "nullable": true
                  },
                  "skills": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "技能名（我的技能 + 开关型内置技能：章节规划/读者与题材/叙事声音）。开关型未列出不注入；基础设施内置（故事状态/成稿质量门）不受此列表控制，始终注入（DELTA-004）"
                  }
                },
                "required": [
                  "content"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/v1/novels/{id}/chapters/chat": {
      "post": {
        "summary": "章节对话（SSE 流式续写/问答，按用户限流）",
        "operationId": "chapterChat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          },
          {
            "name": "chapterId",
            "in": "query",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": { "type": "string" },
                  "skills": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "技能名（场景技能 + 我的技能 + 开关型内置：章节规划/读者与题材/叙事声音）。开关型未列出不注入；基础设施内置（故事状态/成稿质量门）始终注入（DELTA-004）"
                  },
                  "styleId": { "type": "integer", "nullable": true },
                  "generationKey": { "type": "string" }
                },
                "required": ["content"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE 事件流（start/phase/delta/error/done 事件序）",
            "content": { "text/event-stream": { "schema": { "type": "string" } } }
          },
          "400": {
            "description": "参数缺失/非法",
            "content": { "application/json": { "schema": { "type": "object" } } }
          },
          "404": {
            "description": "章节不存在或不归属当前用户",
            "content": { "application/json": { "schema": { "type": "object" } } }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
    "/api/v1/books/{bookId}/export": {
      "get": {
        "summary": "导出作品规范JSON附件",
        "operationId": "exportBook",
        "parameters": [
          {"name": "bookId", "in": "path", "required": true, "schema": { "type": "integer" }},
          {"name": "format", "in": "query", "required": true, "schema": { "type": "string（仅 json）" }}
        ],
        "responses": {
          "200": {
            "description": "application/json attachment：{schemaVersion:'mozhou.novel.export.v1', novel, chapters[], characters[], worldviews[]}"
          },
          "400": {
            "description": "format≠json 或编号无效"
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "作品不存在或不属于当前用户"
          }
        }
      }
    },
...
    "/api/v1/market/briefings": {
      "post": {
        "summary": "生成版本化市场简报产物",
        "operationId": "createMarketBriefing",
        "responses": {
          "201": {
            "description": "{briefing:{artifactId, version}}"
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "暂无市场数据"
          }
        }
      }
    },
...
    "/api/v1/novels/{id}/chapters/chat/stop": {
      "post": {
        "summary": "请求停止章节生成",
        "operationId": "stopChapterGeneration",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": { "type": "integer" }},
          {"name": "chapterId", "in": "query", "required": true, "schema": { "type": "integer" }}
        ],
        "responses": {
          "200": {
            "description": "{stopped:boolean}；命中时补写 cancelled SSE 事件并按 cancelled 记账"
          },
          "400": {
            "description": "参数非法"
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "章节不存在或不属于当前用户"
          }
        }
      }
    },
    "/api/v1/novels/{id}/chapters/messages/{messageId}/discard": {
      "post": {
        "summary": "丢弃AI候选回复（不改正文）",
        "operationId": "discardChapterMessage",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": { "type": "integer" }},
          {"name": "messageId", "in": "path", "required": true, "schema": { "type": "integer" }},
          {"name": "chapterId", "in": "query", "required": true, "schema": { "type": "integer" }}
        ],
        "responses": {
          "200": {
            "description": "{ok:true, status:'discarded'}"
          },
          "400": {
            "description": "参数缺失"
          },
          "401": {
            "description": "未登录"
          },
          "404": {
            "description": "候选不存在/不归属/状态不可丢弃"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "RateLimited": {
        "description": "Too Many Requests：触发限流；Retry-After 头为建议等待秒数。桶与阈值见 api-contract.md §25",
        "content": { "application/json": { "schema": { "type": "object" } } }
      }
    },
...
  }
}